Conversion
We imported math so that we could access the code we need. We also worked on it on the same computer, we put an input so the user could type the decimal number they want to convert. Then we switched it to a binary number and we used that number to convert it to a hexadecimal. We then printed both the binary and hexadecimal number.